ajuste na deleção de cliente

+++ src/main/java/io/github/eduardoconceicao90/desafio_mv/service/ClienteService.java
@@ -6,8 +6,11 @@ import io.github.eduardoconceicao90.desafio_mv.domain.cliente.PessoaJuridica;
 import io.github.eduardoconceicao90.desafio_mv.domain.cliente.dto.ClienteDTO;
 import io.github.eduardoconceicao90.desafio_mv.domain.cliente.dto.PessoaFisicaDTO;
 import io.github.eduardoconceicao90.desafio_mv.domain.cliente.dto.PessoaJuridicaDTO;
+import io.github.eduardoconceicao90.desafio_mv.domain.cliente.enums.StatusCliente;
 import io.github.eduardoconceicao90.desafio_mv.domain.cliente.enums.TipoCliente;
 import io.github.eduardoconceicao90.desafio_mv.repository.ClienteRepository;
+import io.github.eduardoconceicao90.desafio_mv.repository.ContaPFRepository;
+import io.github.eduardoconceicao90.desafio_mv.repository.ContaPJRepository;
 import io.github.eduardoconceicao90.desafio_mv.service.exception.ObjectNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,6 +29,12 @@ public class ClienteService {
     @Autowired
     private ClienteRepository clienteRepository;
 
+    @Autowired
+    private ContaPFRepository contaPFRepository;
+
+    @Autowired
+    private ContaPJRepository contaPJRepository;
+
     public Cliente findById(Long id) {
         Optional<Cliente> cliente = clienteRepository.findById(id);
         return cliente.orElseThrow(() -> new ObjectNotFoundException("Cliente não encontrado! Id: " + id ));
@@ -41,6 +50,7 @@ public class ClienteService {
         var cliente = new ClienteDTO(null, pessoa.getCpf(), null);
         findByDocumento(cliente);
 
+        pessoa.setStatusCliente(StatusCliente.ATIVO);
         pessoa.setTipoCliente(TipoCliente.PESSOA_FISICA);
         return clienteRepository.save(pessoa);
     }
@@ -63,6 +73,7 @@ public class ClienteService {
         var cliente = new ClienteDTO(null, null, pessoa.getCnpj());
         findByDocumento(cliente);
 
+        pessoa.setStatusCliente(StatusCliente.ATIVO);
         pessoa.setTipoCliente(TipoCliente.PESSOA_JURIDICA);
         return clienteRepository.save(pessoa);
     }
@@ -79,10 +90,33 @@ public class ClienteService {
         return clienteRepository.save(pessoaSalva);
     }
 
-    public void deletarCliente(Long id) {
-        findById(id);
-        clienteRepository.deleteById(id);
+    public void deletarOuInativarCliente(Long id) {
+       Cliente cliente = findById(id);
+
+        if(cliente.getTipoCliente().name().equals("PESSOA_FISICA")) {
+
+            String statusConta = contaPFRepository.statusConta(cliente.getId());
+
+            if(statusConta != null){
+                clienteRepository.inativarCliente(id);
+            }else {
+                clienteRepository.deleteById(id);
+            }
+
+        } else if(cliente.getTipoCliente().name().equals("PESSOA_JURIDICA")) {
+
+            String statusConta = contaPJRepository.statusConta(cliente.getId());
+
+            if (statusConta != null) {
+                clienteRepository.inativarCliente(id);
+            }else {
+                clienteRepository.deleteById(id);
+            }
+
+        }
+
     }
+
 
     //------------------------------------------
